% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt-class.R
\name{is_obt}
\alias{is_obt}
\title{Test whether an object is a recipe}
\usage{
is_obt(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} when \code{x} is an \code{obt} recipe, \code{FALSE} otherwise.
}
\description{
Test whether an object is a recipe
}
\examples{
is_obt(obt(folder = tempdir()))
is_obt("a folder path")
}
