% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linelist-class.R
\name{is_obt_linelist}
\alias{is_obt_linelist}
\title{Test whether an object is a linelist recipe}
\usage{
is_obt_linelist(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} when \code{x} is an \code{obt_linelist} recipe, \code{FALSE} otherwise.
}
\description{
Test whether an object is a linelist recipe
}
\examples{
is_obt_linelist(obt_linelist(obt(folder = tempdir())))
is_obt_linelist(obt(folder = tempdir()))
}
