% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-class.R
\name{is_obt_setup}
\alias{is_obt_setup}
\title{Test whether an object is a setup recipe}
\usage{
is_obt_setup(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} when \code{x} is an \code{obt_setup} recipe, \code{FALSE} otherwise.
}
\description{
Test whether an object is a setup recipe
}
\examples{
is_obt_setup(obt_setup(obt(folder = tempdir())))
is_obt_setup(obt(folder = tempdir()))
}
