% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt-package.R
\docType{package}
\name{obt-package}
\alias{obt-package}
\title{obt: Drive the OutbreakTools Designer and Linelist from R}
\description{
An R wrapper around the OutbreakTools designer and the linelist workbook buttons, so the steps normally driven from the Excel interface can be scripted from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epicentre-msf/obt}
  \item Report bugs at \url{https://github.com/epicentre-msf/obt/issues}
}

}
\author{
\strong{Maintainer}: Komlavi Yves Amevoin \email{Yves.Amevoin@akyves.net}

Authors:
\itemize{
  \item Komlavi Yves Amevoin \email{Yves.Amevoin@akyves.net}
}

}
\keyword{internal}
