% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_recode_letter_percent.R
\name{obt_recode_letter_percent}
\alias{obt_recode_letter_percent}
\title{Recode Letter to Percent}
\usage{
obt_recode_letter_percent(x)
}
\arguments{
\item{x}{A vector of report card marks containing alpha marks, or a combination of alpha and mumeric marks.}
}
\value{
Numeric vector
}
\description{
A function to recode alpha report card marks (A, B-, D+) to Numeric values.
}
\examples{
# ele_rc$reading_numeric <- obt_recode_letter_percent(ele_rc$reading_numeric)
}
