% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_recode_rc_pass_fail.R
\name{obt_recode_rc_pass_fail}
\alias{obt_recode_rc_pass_fail}
\title{Recode Report Card marks to Pass/Fail}
\usage{
obt_recode_rc_pass_fail(x)
}
\arguments{
\item{x}{}
}
\value{
Character vector with Pass/Fail
}
\description{
Recode Report Card marks to Pass/Fail
}
\examples{
df <- ele_rc
ele_rc$reading
[1] "C+" "A-" "C"  "C"  "B-" "C-" "B-" "A-" "78" "D"  "R"  "B-" "60" "D"  "B-"

obt_recode_rc_pass_fail(ele_rc$reading)
[1] "pass" "pass" "pass" "pass" "pass" "pass" "pass" "pass" "pass" "pass" "fail" "pass" "pass" "pass" "pass"

}
