% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_sec_level_study.R
\name{obt_sec_level_study}
\alias{obt_sec_level_study}
\title{Secondary Level of Study}
\usage{
obt_sec_level_study(x)
}
\arguments{
\item{x}{A vector containing an Ontario secondary course code with at least the first five digits (e.g. ENG2D, FSF3U, SNC1L)}
}
\value{
A vector containing the course level for associated course code (e.g. ENG2D = Academic, FSF3U = University)
}
\description{
A function to re-code Ontario secondary courses into Level of Study (Academic, Applied, University, College, Open etc.)
}
\examples{
# obt_sec_level_study(sec_rc$course_number)
}
