% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_summary_table.R
\name{obt_summary_table}
\alias{obt_summary_table}
\title{Summary Table}
\usage{
obt_summary_table(df, group_1, group_2)
}
\arguments{
\item{df}{Dataframe or tibble}

\item{group_1}{First grouping variable}

\item{group_2}{Second grouping variable (optional)}
}
\value{
A tibble with counts and percentages of each group
}
\description{
Given a dataframe or tibble, returns a tibble with counts and percentages
of a group. Can be a single group, or group 2 can be nested within group 1
}
\examples{
obt_summary_table (mtcars, gear)
obt_summary_table (mtcars, gear, carb)
}
