% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt-package.R
\docType{package}
\name{obt-package}
\alias{obt}
\alias{obt-package}
\title{obt: Functions for Ontario School Board data professionals}
\description{
The obt (Ontario Board Tools) is a collection of functions for data professionals working in Ontario school boards. These functions are designed to assist those working with common data sets across the province, such as Report Card, EQAO, OUAC, OCAS and others.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/grousell/obt}
  \item Report bugs at \url{https://github.com/grousell/obt/issues}
}

}
\author{
\strong{Maintainer}: Rousell Greg \email{grousell@gmail.com}

}
\keyword{internal}
