% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_recode_rc_L34.R
\name{obt_recode_rc_L34}
\alias{obt_recode_rc_L34}
\title{Recode Report Card marks to Met Standard / Did not Meet Standard}
\usage{
obt_recode_rc_L34(x)
}
\arguments{
\item{x}{}
}
\value{
Character vector with Met Standard / Did not Meet Standard
}
\description{
Recode Report Card marks to Met Standard / Did not Meet Standard
}
