% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_recode_rc_pass_fail.R
\name{obt_recode_rc_pass_fail}
\alias{obt_recode_rc_pass_fail}
\title{Recode Report Card marks to Pass/Fail}
\usage{
obt_recode_rc_pass_fail(x)
}
\arguments{
\item{x}{}
}
\value{
Character vector with Pass/Fail
}
\description{
All alpha and numeric marks are re-coded to either pass or fail, including R.
I remains the same as it's "Insufficient Evidence"
}
\examples{
df <- ele_rc
ele_rc$reading
[1] "R"  "I"  "D"  "C"  "A-"

obt_recode_rc_pass_fail(ele_rc$reading)
[1] "fail" "I" "pass" "pass" "pass"

}
