% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_recode_rc_to_levels.R
\name{obt_recode_rc_to_levels}
\alias{obt_recode_rc_to_levels}
\title{Recode Report Card Marks to Achievement Levels}
\usage{
obt_recode_rc_to_levels(x)
}
\arguments{
\item{x}{Vector of report card marks. Can be either alpha (C+, B-, A) or numeric (88, 63, 79)}
}
\value{
Factor vector with achievement levels (NE1, Level 1, Level 2, Level 3, Level 4)
}
\description{
This function takes report card marks, both alpha (A+, B, C-, etc.) and numeric (68, 72, 86, etc.)
and converts to achievement levels used in Ontario school boards
}
