% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obt_summary_table.R
\name{obt_summary_table}
\alias{obt_summary_table}
\title{Summary Table}
\usage{
obt_summary_table(df, group_1, group_2, ...)
}
\arguments{
\item{df}{}

\item{group_1}{}

\item{group_2}{}

\item{...}{}
}
\value{
A tibble with counts and percentages of each group
}
\description{
Given a dataframe or tibble, returns a tibble with counts and percentages
of a group. Can be a single group, or group 2 can be nested within group 1
}
\examples{
dat <- palmerpenguins::penguins
obt_summary_table (dat, species)
obt_summary_table (dat, species, sex)
}
